+++ eyes.sdk.core/src/main/java/com/applitools/ICheckSettingsInternal.java
@@ -2,8 +2,10 @@ package com.applitools;
 
 import com.applitools.eyes.MatchLevel;
 import com.applitools.eyes.Region;
+import com.applitools.eyes.visualgrid.model.VisualGridOption;
 import com.applitools.eyes.visualgrid.model.VisualGridSelector;
 
+import java.util.List;
 import java.util.Map;
 
 public interface ICheckSettingsInternal {
@@ -31,4 +33,6 @@ public interface ICheckSettingsInternal {
     Boolean isStitchContent();
 
     Boolean isIgnoreDisplacements();
+
+    List<VisualGridOption> getVisualGridOptions();
 }