+++ eyes.sdk.core/src/main/java/com/applitools/eyes/visualgrid/model/RenderingTask.java
@@ -430,7 +430,8 @@ public class RenderingTask implements Callable<RenderStatusResults>, Completable
 
             RenderRequest request = new RenderRequest(this.renderingInfo.getResultsUrl(), result.getUrl(), dom,
                     resourceMapping, renderInfo, browserInfo.getPlatform(), browserInfo.getBrowserType(),
-                    checkSettingsInternal.getScriptHooks(), regionSelectorsList, checkSettingsInternal.isSendDom(), visualGridTask, this.renderingInfo.getStitchingServiceUrl());
+                    checkSettingsInternal.getScriptHooks(), regionSelectorsList, checkSettingsInternal.isSendDom(),
+                    visualGridTask, this.renderingInfo.getStitchingServiceUrl(), checkSettingsInternal.getVisualGridOptions());
 
             allRequestsForRG.add(request);
         }